requirements.txt
colorama==0.4.6
pygame==2.5.2
Pillow==10.1.0
